update client.
Now leaderBoard work.
Score = Enemy.exp * 10;
When game end, update score if score is higher than highscore

+++ Spring/LoginSystem/src/main/java/kr/ac/ajou/mse/loginsystem/ProducedUserData.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 public class ProducedUserData {
     private String uuid;
     private String name;
-    private float highscore;
+    private long highscore;
     private long scoreid;
 
     public ProducedUserData (UserData u) {
